third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/disable-breakpoints-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests disabling breakpoints.


Running: testSetBreakpointPauseResumeThenDisable
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (disable-breakpoints.js:13)
Breakpoint sidebar pane 
disable-breakpoints.js:13          var x = Math.sqrt(10);
Script execution resumed.
Test function finished.
Disabling breakpoints...
Running test function again...
Set timer for test function.
Test function finished.

Running: testEnableBreakpointsAgain
Enabling breakpoints...
Running test function...
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (disable-breakpoints.js:13)
Breakpoint sidebar pane 
disable-breakpoints.js:13          var x = Math.sqrt(10);
Script execution resumed.
Test function finished.
Breakpoint sidebar pane 
disable-breakpoints.js:13          var x = Math.sqrt(10);
Breakpoints removed.
Breakpoint sidebar pane 
No breakpoints
